use X509CertRef to get the namespace of the cert
X509CertRef is a ResourceRefNamespaced, which means it contains a name and a namespace. The name is only used to get the object while the namespace is discarded, This PR uses the namespace defined here to get the object while falling back to the project namespace to maintain backward compatibility.

This can break projects for people configuring the namespace differently that what's intended. It will only break if the configuration is initially wrong.