CLOUDP-321083: fix reconciliation loop on dbuser (#2376)
* CLOUDP-321083: fix reconciliation loop on dbuser

In 140ba21f32fd5c77c2b070b8851b7a7b23a4b2f7 I added the ability to set a
description on AtlasDatabaseUser. However, I overlooked that a
description could already exist on the user in Atlas, in which case the
operator will fetch it, realize it's different, and then... omit the
change, because `omitempty` is set on the struct field.

With this patch we send an empty string instead.

* Add tests