CLOUDP-330314: Disallow project/deployment renaming
When a project is renamed in Kube, the operator will create a new
project and leave the original intact, while other resources (such as
AtlasDeployment) still track the original clusters in the old project.
This leads to bad situations that are difficult to clean up, in
particular via the operator.

Similarly, when a deployment is renamed in Kube a new deployment is
created in Atlas, with the old one left running. Since this is almost
definitely not what the user wanted, let's ensure that name changes here
are also forbidden.