Disallow project renaming
While this is possible in Atlas, the implementation in the operator
causes a new project to be created. Other resources like AtlasDeployment
then continue to track the clusters in the original project.

Since that behavior is clearly wrong, let's forbid changing this field
until AKO can implement renaming support for projects properly.